requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
parsel==1.6.0
Pillow==8.0.1
Protego==0.1.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==2.0.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pytz==2020.4
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
scrapy-djangoitem==1.1.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
urllib3==1.26.2
w3lib==1.22.0
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.2.0
